templates/fastapi/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
greenlet>=3.0.1
python-dotenv>=1.0.0
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9  # for PostgreSQL support
email-validator>=2.1.0  # explicit requirement for email validation
